// donation-svelte-0.8.0-ts-v3/src/services/donation-types.ts
export interface User {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  _id: string;
}

export interface LoggedInUser {
  email: string;
  token: string;
  _id: string;
}

export interface Donor {
  firstName: string;
  lastName: string;
}

export interface Candidate {
  firstName: string;
  lastName: string;
  office: string;
  _id: string;
}

export interface Donation {
  amount: number;
  method: string;
  candidate: Candidate;
  donor: Donor;
  lat: number;
  lng: number;
  _id: string;
}

export interface CandidateDonations {
  candidate: Candidate;
  donations: Donation[];
}

export interface ChartData {
  labels: string[];
  datasets: [{ values: number[] }];
}